refactor: change method dp2Px return int type as float

+++ app/src/main/java/org/getbuddies/a2step/ui/extendz/DisplayExtends.kt
@@ -3,12 +3,12 @@ package org.getbuddies.a2step.ui.extendz
 import android.content.Context
 import org.getbuddies.a2step.App
 
-fun Float.dpToPx(): Int {
+fun Float.dpToPx(): Float {
     val scale: Float = App.get().resources.displayMetrics.density
-    return (this * scale + 0.5f).toInt()
+    return this * scale + 0.5f
 }
 
-fun Float.pxToDp(): Int {
+fun Float.pxToDp(): Float {
     val scale: Float = App.get().resources.displayMetrics.density
-    return (this / scale + 0.5f).toInt()
+    return this / scale + 0.5f
 }
